+++ src/lib.rs
@@ -11,11 +11,13 @@
 //! # use buffer_graphics_lib::drawable::{Drawable, stroke};
 //! # use buffer_graphics_lib::Graphics;
 //! # use buffer_graphics_lib::shapes::CreateDrawable;
+//! # use buffer_graphics_lib::text::pos::TextPos;
+//! # use buffer_graphics_lib::text::pos::NewTextPos;
 //! # use buffer_graphics_lib::text::Text;
 //! # use buffer_graphics_lib::text::TextSize::Large;
 //! let mut buffer = [0_u8; 800 * 600 * 4]; //800 x 600 RGBA
 //! let mut graphics = Graphics::new(&mut buffer, 800, 600).unwrap();
-//! let text = Text::new("Some text", (1,1), (LIGHT_GRAY, Large));
+//! let text = Text::new("Some text", TextPos::cr((1,1)), (LIGHT_GRAY, Large));
 //! graphics.draw(&text);
 //! let drawable = Drawable::from_obj(Rect::new((1,1),(15,16)), stroke(BLUE));
 //! graphics.draw(&drawable);