+++ src/text/pos.rs
@@ -19,21 +19,23 @@ impl TextPos {
     }
 }
 
-macro_rules! impl_to_px {
-    ($num_type: ty) => {
-        impl From<($num_type, $num_type)> for TextPos {
-            fn from((x, y): ($num_type, $num_type)) -> Self {
-                TextPos::Px(x as isize, y as isize)
-            }
-        }
-    };
+pub trait NewTextPos<T> {
+    /// Creates a new TextPos::ColRow
+    fn cr(xy: (T, T)) -> TextPos;
+    /// Creates a new TextPos::Px
+    fn px(xy: (T, T)) -> TextPos;
 }
 
-macro_rules! impl_to_colrow {
+macro_rules! impl_to_px {
     ($num_type: ty) => {
-        impl From<($num_type, $num_type)> for TextPos {
-            fn from((x, y): ($num_type, $num_type)) -> Self {
-                TextPos::ColRow(x as usize, y as usize)
+        impl NewTextPos<$num_type> for TextPos {
+            /// Calls abs() on x and y
+            fn cr((x, y): ($num_type, $num_type)) -> Self {
+                TextPos::ColRow(((x as isize).abs()) as usize, ((y as isize).abs()) as usize)
+            }
+
+            fn px((x, y): ($num_type, $num_type)) -> Self {
+                TextPos::Px(x as isize, y as isize)
             }
         }
     };
@@ -45,20 +47,11 @@ impl_to_px!(i16);
 impl_to_px!(i32);
 impl_to_px!(i64);
 impl_to_px!(i128);
-
-impl_to_colrow!(usize);
-impl_to_colrow!(u8);
-impl_to_colrow!(u16);
-impl_to_colrow!(u32);
-impl_to_colrow!(u64);
-impl_to_colrow!(u128);
-
-impl TextPos {
-    pub fn usize_px(x: usize, y: usize) -> TextPos {
-        TextPos::Px(x as isize, y as isize)
-    }
-
-    pub fn f32_px(x: f32, y: f32) -> TextPos {
-        TextPos::Px(x as isize, y as isize)
-    }
-}
+impl_to_px!(usize);
+impl_to_px!(u8);
+impl_to_px!(u16);
+impl_to_px!(u32);
+impl_to_px!(u64);
+impl_to_px!(u128);
+impl_to_px!(f32);
+impl_to_px!(f64);